move DescriptorFormat test to tests/utils
